+++ Assets/Audio/TestAssets/The Truth Hurts at The Stardrop Saloon.wav.meta
