+++ Assets/Scenes/Act1/Chapter1.unity
@@ -182,7 +182,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   hoverTextColor: {r: 1, g: 1, b: 1, a: 1}
-  _tmp: {fileID: 0}
 --- !u!114 &166288183
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -196,15 +195,15 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_Navigation:
-    m_Mode: 3
-    m_SelectOnUp: {fileID: 0}
-    m_SelectOnDown: {fileID: 0}
-    m_SelectOnLeft: {fileID: 0}
-    m_SelectOnRight: {fileID: 0}
+    m_Mode: 4
+    m_SelectOnUp: {fileID: 166288183}
+    m_SelectOnDown: {fileID: 737597468}
+    m_SelectOnLeft: {fileID: 166288183}
+    m_SelectOnRight: {fileID: 737597468}
   m_Transition: 1
   m_Colors:
     m_NormalColor: {r: 1, g: 1, b: 1, a: 1}
-    m_HighlightedColor: {r: 0.65882355, g: 0.3254902, b: 0.4392157, a: 1}
+    m_HighlightedColor: {r: 1, g: 1, b: 1, a: 1}
     m_PressedColor: {r: 0.65882355, g: 0.3254902, b: 0.4392157, a: 1}
     m_SelectedColor: {r: 0.65882355, g: 0.3254902, b: 0.4392157, a: 1}
     m_DisabledColor: {r: 0.78431374, g: 0.78431374, b: 0.78431374, a: 0.5019608}
@@ -411,18 +410,7 @@ MonoBehaviour:
         m_CallState: 2
   onLineStart:
     m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 1483509596}
-        m_MethodName: playCharacterSound
-        m_Mode: 1
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
+      m_Calls: []
   onLineFinishDisplaying:
     m_PersistentCalls:
       m_Calls: []
@@ -618,7 +606,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   hoverTextColor: {r: 1, g: 1, b: 1, a: 1}
-  _tmp: {fileID: 0}
 --- !u!114 &737597468
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -632,15 +619,15 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_Navigation:
-    m_Mode: 3
-    m_SelectOnUp: {fileID: 0}
-    m_SelectOnDown: {fileID: 0}
-    m_SelectOnLeft: {fileID: 0}
-    m_SelectOnRight: {fileID: 0}
+    m_Mode: 4
+    m_SelectOnUp: {fileID: 166288183}
+    m_SelectOnDown: {fileID: 737597468}
+    m_SelectOnLeft: {fileID: 166288183}
+    m_SelectOnRight: {fileID: 737597468}
   m_Transition: 1
   m_Colors:
     m_NormalColor: {r: 1, g: 1, b: 1, a: 1}
-    m_HighlightedColor: {r: 0.65882355, g: 0.3254902, b: 0.4392157, a: 1}
+    m_HighlightedColor: {r: 1, g: 1, b: 1, a: 1}
     m_PressedColor: {r: 0.65882355, g: 0.3254902, b: 0.4392157, a: 1}
     m_SelectedColor: {r: 0.65882355, g: 0.3254902, b: 0.4392157, a: 1}
     m_DisabledColor: {r: 0.78431374, g: 0.78431374, b: 0.78431374, a: 0.5019608}
@@ -1442,9 +1429,9 @@ RectTransform:
   m_Father: {fileID: 166288181}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 160, y: 0}
+  m_AnchorMin: {x: 0, y: 1}
+  m_AnchorMax: {x: 0, y: 1}
+  m_AnchoredPosition: {x: 160, y: -102.005}
   m_SizeDelta: {x: 260.64, y: 174.01}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1783023151
